move scripts in fpga and os dirs
add os/systemd, move tests in drivers
