+++ circle.yml
@@ -7,4 +7,4 @@ dependencies:
     - docker build -t koheron/zynq-sdk .
 test:
   pre:
-    - docker run --env "DOCKER=True" koheron/zynq-sdk  bash scripts/ci_tests.sh
+    - docker run --env "DOCKER=True" koheron/zynq-sdk  bash scripts/tests/ci_tests.sh
