+++ fpga/lib/cfg_sts.tcl
