+++ fpga/lib/starting_point.tcl
