+++ fpga/scripts/hwdef.tcl
