+++ fpga/scripts/test_core.tcl
