+++ os/tests/instrument_manager.py
