+++ os/tests/tests_instrument_manager.py
