--- tests/run_tests.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-set -e
-
-py.test -v tests_instrument_manager.py
-py.test -v tests_device_memory.py
